requirements.txt
alembic==1.12.0
Babel==2.14.0
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.13.0
cachetools==5.3.3
captcha==0.5.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.7
dnspython==2.4.2
email-validator==2.1.0.post1
filelock==3.13.1
Flask==2.3.3
Flask-Admin==1.6.1
flask-babel
Flask-BasicAuth==0.2.0
Flask-Bcrypt==1.0.1
flask-change-password==0.0.7
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-MySQL==1.5.2
Flask-MySQLdb==2.0.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-Session==0.8.0
Flask-Sessionstore==0.4.5
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
flask_session_captcha==1.4.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
infinity==1.5
iniconfig==2.0.0
install==1.3.5
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
msgspec==0.18.6
mysqlclient==2.2.1
nose==1.3.7
oauthlib==3.2.2
optional-django==0.3.0
packaging==24.0
passlib==1.7.4
Pillow==10.0.0
platformdirs==3.11.0
pluggy==1.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pymongo==4.7.2
PyMySQL==1.1.0
pytest==8.2.0
python-dotenv==1.0.1
pytz==2023.3.post1
react==4.3.0
redis==5.0.4
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.20
typing_extensions==4.7.1
urllib3==2.1.0
validators==0.22.0
virtualenv==20.24.6
Werkzeug==2.3.7
WTForms==3.0.1
WTForms-Components==0.10.5
gunicorn
itsdangerous